+++ restapi/api/basic_endpoints.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"github.com/danielpaulus/go-ios/ios"
-	"github.com/danielpaulus/go-ios/ios/screenshotr"
 	"github.com/gin-gonic/gin"
 	log "github.com/sirupsen/logrus"
 	"io"
@@ -28,15 +27,4 @@ func List(c *gin.Context) {
 	a, _, _ := ios.Listen()
 	list, _ := a()
 	c.IndentedJSON(http.StatusOK, list)
-}
-
-func Screenshot(c *gin.Context) {
-
-	dev, _ := ios.GetDevice("")
-	conn, err := screenshotr.New(dev)
-	log.Error(err)
-	b, _ := conn.TakeScreenshot()
-
-	c.Header("Content-Type", "image/png")
-	c.Data(http.StatusOK, "application/octet-stream", b)
 }
